First commit
Template creation.
The "crear_post" form was created.
"Index" loads the posts and has its own CSS. The navigation bar is in Layouts and now includes a "Home" button that returns the user to the posts; it also has its own CSS.
Posts can be accessed using the HTML tag "<a>", taking the user to the corresponding post with its replies.
The "base_post" template is the literal basis for how posts are loaded when accessed.
The "tablero" app has its own models. Besides "Users" (not used) and "posts," there is "responses," used to store post replies. A foreign key is used to link them to the corresponding post.